requirements.txt
cbor2==5.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
docopt==0.6.2
idna==3.11
lxml==6.0.4
pipreqs==0.4.13
pycparser==3.0
pycryptodome==3.23.0
pyscard==2.3.1
python-docx==1.2.0
pywin32==311
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
yarg==0.1.10
